requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==39.0.1
dj-database-url==1.2.0
Django==4.1.6
django-extensions==3.2.1
django-filter==22.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
gunicorn==20.1.0
inflection==0.5.1
psycopg2-binary==2.9.5
pycparser==2.21
python-dotenv==0.21.1
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.3.0
pandas==1.3.3
numpy==1.21.2
matplotlib==3.4.3